+++ src/happy/readers/_envi_reader.py
@@ -4,6 +4,7 @@ import re
 from ._happydata_reader import HappyDataReader
 from happy.readers.spectra import EnviReader
 from happy.data import HappyData
+from seppl.placeholders import expand_placeholders
 
 from typing import List, Optional
 
@@ -36,8 +37,9 @@ class EnviHappyDataReader(HappyDataReader):
 
     def _get_sample_ids(self) -> List[str]:
         sample_ids = []
-        for sample_id in os.listdir(self.base_dir):
-            sample_path = os.path.join(self.base_dir, sample_id)
+        base_dir = expand_placeholders(self.base_dir)
+        for sample_id in os.listdir(base_dir):
+            sample_path = os.path.join(base_dir, sample_id)
             if sample_id.endswith(self.extension) and os.path.isfile(sample_path):
                 add = True
                 if self.exclude is not None:
@@ -59,13 +61,14 @@ class EnviHappyDataReader(HappyDataReader):
         def filename_func(base_dir, sample_id):
             return os.path.join(base_dir, sample_id + self.extension)
 
-        hyperspec_file_path = os.path.join(self.base_dir, sample_id + region_name + self.extension)
+        base_dir = expand_placeholders(self.base_dir)
+        hyperspec_file_path = os.path.join(base_dir, sample_id + region_name + self.extension)
         self.logger().info(f"{sample_id}{region_name}{self.extension}")
 
         if not os.path.exists(hyperspec_file_path):
             raise ValueError(f"Hyperspectral ENVI file not found for sample_id: {sample_id}")
 
-        envi_reader = EnviReader(self.base_dir, filename_func=filename_func)
+        envi_reader = EnviReader(base_dir, filename_func=filename_func)
         envi_reader.load_data(sample_id)
         hyperspec_data = envi_reader.get_numpy()