+++ src/happy/readers/_happy_reader.py
@@ -6,6 +6,7 @@ from ._happydata_reader import HappyDataReader
 from happy.readers.spectra import EnviReader
 from happy.data import HappyData, MASK_MAP
 import spectral.io.envi as envi
+from seppl.placeholders import expand_placeholders
 
 from typing import List, Optional, Tuple, Dict
 
@@ -116,8 +117,9 @@ class HappyReader(HappyDataReader):
 
     def _get_sample_ids(self) -> List[str]:
         sample_ids = []
-        for sample_id in os.listdir(self.base_dir):
-            sample_path = os.path.join(self.base_dir, sample_id)
+        base_dir = expand_placeholders(self.base_dir)
+        for sample_id in os.listdir(base_dir):
+            sample_path = os.path.join(base_dir, sample_id)
             if os.path.isdir(sample_path):
                 sample_ids.append(sample_id)
         sample_ids = sorted(sample_ids)
@@ -149,9 +151,10 @@ class HappyReader(HappyDataReader):
         return happy_data_list
 
     def _load_region(self, sample_id: str, region_name: str) -> HappyData:
-        hyperspec_file_path = os.path.join(self.base_dir, sample_id, region_name, f'{sample_id}'+".hdr")
+        base_dir = expand_placeholders(self.base_dir)
+        hyperspec_file_path = os.path.join(base_dir, sample_id, region_name, f'{sample_id}'+".hdr")
         self.logger().info(f"{sample_id}:{region_name}")
-        base_path = os.path.join(self.base_dir, sample_id, region_name)
+        base_path = os.path.join(base_dir, sample_id, region_name)
 
         if not os.path.exists(hyperspec_file_path):
             raise ValueError(f"Hyperspectral ENVI file not found for sample_id: {sample_id}")
@@ -173,7 +176,7 @@ class HappyReader(HappyDataReader):
             metadata_dict[target_name] = {"data": metadata, "mapping": mapping}
 
         # mask.json
-        mask_map = self._load_json(os.path.join(self.base_dir, sample_id, region_name, "mask.json"))
+        mask_map = self._load_json(os.path.join(base_dir, sample_id, region_name, "mask.json"))
         if len(mask_map) > 0:
             metadata_dict[MASK_MAP] = mask_map
 
@@ -215,7 +218,7 @@ class HappyReader(HappyDataReader):
         return metadata_file_paths
 
     def _get_regions(self, sample_id: str) -> List[str]:
-        sample_dir = os.path.join(self.base_dir, sample_id)
+        sample_dir = os.path.join(expand_placeholders(self.base_dir), sample_id)
         if os.path.isdir(sample_dir):
             region_dirs = [name for name in os.listdir(sample_dir) if os.path.isdir(os.path.join(sample_dir, name))]
         else: